backend/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
pydantic>=2.0.0
google-generativeai>=0.3.0
APScheduler==3.10.4
pytz==2024.1
